modules/nllb/requirements.txt
runpod==0.10.0
diffusers==0.17.1
transformers==4.30.2
safetensors==0.3.1
ftfy==6.1.1
scipy==1.9.0
accelerate==0.18.0
bitsandbytes==0.37.2
numpy==1.24.1
opencv-python==4.7.0.68
tqdm==4.65.0
Pillow==9.5.0
timm==0.4.12
ipython==7.19.0
huggingface-hub==0.15.1
boto3>=1.26.63,<2
boto3-type-annotations>=0.3.1,<1
responses==0.22.0
httpx==0.23.1
hypothesis==6.56.4
redis>=4,<5
python-dotenv
tabulate==0.9.0
pydantic==1.10.5
flask==2.2.3
waitress==2.1.2
nltk==3.8.1
lingua-language-detector==1.3.2
triton==2.0.0